refactor: extract SalesCard into its own file
